+++ bom/bom-quarkus/pom.xml
@@ -9,18 +9,12 @@
 	<artifactId>bom-quarkus</artifactId>
 	<packaging>pom</packaging>
 
-	<properties>
-		<quarkus.platform.artifact-id>quarkus-bom</quarkus.platform.artifact-id>
-		<quarkus.platform.group-id>io.quarkus.platform</quarkus.platform.group-id>
-		<quarkus.platform.version>2.11.1.Final</quarkus.platform.version>
-	</properties>
-
 	<dependencyManagement>
 		<dependencies>
 			<dependency>
-				<groupId>${quarkus.platform.group-id}</groupId>
-				<artifactId>${quarkus.platform.artifact-id}</artifactId>
-				<version>${quarkus.platform.version}</version>
+				<groupId>io.quarkus.platform</groupId>
+				<artifactId>quarkus-bom</artifactId>
+				<version>2.11.1.Final</version>
 				<type>pom</type>
 				<scope>import</scope>
 			</dependency>
