--- sample-quarkus/sample-quarkus-rest-service/src/main/java/hu/icellmobilsoft/sampler/sample/quarkus/restservice/rest/jsonb/FieldOnlyVisibilityStrategy.java
@@ -1,46 +1,0 @@
-/*-
- * #%L
- * Sampler
- * %%
- * Copyright (C) 2022 i-Cell Mobilsoft Zrt.
- * %%
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *      http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * #L%
- */
-package hu.icellmobilsoft.sampler.sample.quarkus.restservice.rest.jsonb;
-
-import java.lang.reflect.Field;
-import java.lang.reflect.Method;
-
-import javax.json.bind.config.PropertyVisibilityStrategy;
-
-/**
- * The default JSON-B output mapping has all public methods, include setter and getter. This configuration avoids this. Based on
- * <a href="https://www.adam-bien.com/roller/abien/entry/private_fields_serialization_with_json">private_fields_serialization_with_json</a>
- * 
- * @author imre.scheffer
- * @since 0.1.0
- *
- */
-public class FieldOnlyVisibilityStrategy implements PropertyVisibilityStrategy {
-
-    @Override
-    public boolean isVisible(Field field) {
-        return true;
-    }
-
-    @Override
-    public boolean isVisible(Method method) {
-        return false;
-    }
-}
